refactor: clean up utility functions and generateSwagger
