+++ src/ServerlessAutoSwagger.ts
@@ -5,7 +5,7 @@ import type { Options } from 'serverless';
 import type { Service } from 'serverless/aws';
 import type { Logging } from 'serverless/classes/Plugin';
 import { getOpenApiWriter, getTypeScriptReader, makeConverter } from 'typeconv';
-import { generateEmptySwagger, removeStringFromArray, writeFile } from './helperFunctions';
+import { generateEmptySwagger, writeFile } from './helperFunctions';
 import swaggerFunctions from './resources/functions';
 import * as customPropertiesSchema from './schemas/custom-properties.schema.json';
 import * as functionEventPropertiesSchema from './schemas/function-event-properties.schema.json';
@@ -15,23 +15,12 @@ import type {
   CustomHttpApiEvent,
   CustomHttpEvent,
   CustomServerless,
-  HeaderParameters,
-  HttpResponses,
-  PathParameterPath,
-  PathParameters,
-  QueryStringParameters,
   ServerlessCommands,
   ServerlessHooks,
 } from './types/serverless-plugin.types';
 
-import type {
-  Definition,
-  MethodSecurity,
-  Parameter,
-  Response,
-  SecurityDefinition,
-  Swagger,
-} from './types/swagger.types';
+import type { Definition, MethodSecurity, SecurityDefinition, Swagger } from './types/swagger.types';
+import { formatResponses, httpEventToParameters } from './converters';
 
 export default class ServerlessAutoSwagger {
   serverless: CustomServerless;
@@ -101,9 +90,20 @@ export default class ServerlessAutoSwagger {
     this.addEndpointsAndLambda();
   };
 
+  generateSwagger = async () => {
+    await this.gatherTypes();
+    this.gatherSwaggerOverrides();
+    this.generateSecurity();
+    this.generatePaths();
+
+    this.log.notice('Creating Swagger file...');
+    const resourcesPath = await this.prepareResourceFolder();
+    await this.writeSwaggerFile(resourcesPath);
+  };
+
   /** Updates this.swagger with serverless custom.autoswagger overrides */
   gatherSwaggerOverrides = (): void => {
-    const autoswagger = this.serverless.service.custom?.autoswagger ?? {};
+    const autoswagger = this.autoSwaggerCustomConfig;
 
     if (autoswagger.basePath) this.swagger.basePath = autoswagger.basePath;
     if (autoswagger.host) this.swagger.host = autoswagger.host;
@@ -201,37 +201,44 @@ export default class ServerlessAutoSwagger {
     //  that may be defined in a custom swagger json
   };
 
-  generateSwagger = async () => {
-    await this.gatherTypes();
-    this.gatherSwaggerOverrides();
-    this.generateSecurity();
-    this.generatePaths();
-
-    this.log.notice('Creating Swagger file...');
-
+  private async prepareResourceFolder() {
     // TODO enable user to specify swagger file path. also needs to update the swagger json endpoint.
     const packagePath = dirname(require.resolve('serverless-auto-swagger/package.json'));
     const resourcesPath = `${packagePath}/dist/resources`;
     await copy(resourcesPath, './swagger');
+    return resourcesPath;
+  }
 
-    if (this.serverless.service.provider.runtime?.includes('python')) {
-      const swaggerStr = JSON.stringify(this.swagger, null, 2)
-        .replace(/true/g, 'True')
-        .replace(/false/g, 'False')
-        .replace(/null/g, 'None');
-      let swaggerPythonString = `# this file was generated by serverless-auto-swagger`;
-      swaggerPythonString += `\ndocs = ${swaggerStr}`;
-      await writeFile('./swagger/swagger.py', swaggerPythonString);
+  private async writeSwaggerFile(resourcesPath: string) {
+    if (this.isPythonRuntime()) {
+      await this.writePythonSwaggerFile();
     } else {
-      await copy(resourcesPath, './swagger', {
-        filter: (src) => src.slice(-2) === 'js',
-      });
-
-      const swaggerJavaScriptString = `// this file was generated by serverless-auto-swagger
-            module.exports = ${JSON.stringify(this.swagger, null, 2)};`;
-      await writeFile('./swagger/swagger.js', swaggerJavaScriptString);
+      await this.writeJSSwaggerFile(resourcesPath);
     }
-  };
+  }
+
+  private isPythonRuntime() {
+    return this.serverless.service.provider.runtime?.includes('python');
+  }
+
+  private async writePythonSwaggerFile() {
+    const swaggerStr = JSON.stringify(this.swagger, null, 2)
+      .replace(/true/g, 'True')
+      .replace(/false/g, 'False')
+      .replace(/null/g, 'None');
+    let swaggerPythonString = `# this file was generated by serverless-auto-swagger`;
+    swaggerPythonString += `\ndocs = ${swaggerStr}`;
+    await writeFile('./swagger/swagger.py', swaggerPythonString);
+  }
+  private async writeJSSwaggerFile(resourcesPath: string) {
+    await copy(resourcesPath, './swagger', {
+      filter: (src) => src.slice(-2) === 'js',
+    });
+
+    const swaggerJavaScriptString = `// this file was generated by serverless-auto-swagger
+            module.exports = ${JSON.stringify(this.swagger, null, 2)};`;
+    await writeFile('./swagger/swagger.js', swaggerJavaScriptString);
+  }
 
   addEndpointsAndLambda = () => {
     this.serverless.service.functions = {
@@ -240,7 +247,7 @@ export default class ServerlessAutoSwagger {
     };
   };
 
-  addSwaggerPath = (functionName: string, http: CustomHttpEvent | CustomHttpApiEvent | string) => {
+  addSwaggerPath = (functionName: string, http: CustomHttpEvent | CustomHttpApiEvent) => {
     if (typeof http === 'string') {
       // TODO they're using the shorthand - parse that into object.
       //  You'll also have to remove the `typeof http !== 'string'` check from the function calling this one
@@ -262,8 +269,8 @@ export default class ServerlessAutoSwagger {
       produces: http.produces ?? ['application/json'],
       security: http.security,
       // This is actually type `HttpEvent | HttpApiEvent`, but we can lie since only HttpEvent params (or shared params) are used
-      parameters: this.httpEventToParameters(http as CustomHttpEvent),
-      responses: this.formatResponses(http.responseData ?? http.responses),
+      parameters: httpEventToParameters(http as CustomHttpEvent),
+      responses: formatResponses(http.responseData ?? http.responses),
     };
 
     const apiKeyHeaders = this.autoSwaggerCustomConfig.apiKeyHeaders;
@@ -290,125 +297,5 @@ export default class ServerlessAutoSwagger {
         .filter((http) => !!http && typeof http !== 'string' && !http.exclude)
         .forEach((http) => this.addSwaggerPath(functionName, http!));
     });
-  };
-
-  formatResponses = (responseData: HttpResponses | undefined) => {
-    if (!responseData) {
-      // could throw error
-      return { 200: { description: '200 response' } };
-    }
-    const formatted: Record<string, Response> = {};
-    Object.entries(responseData).forEach(([statusCode, responseDetails]) => {
-      if (typeof responseDetails == 'string') {
-        formatted[statusCode] = {
-          description: responseDetails,
-        };
-        return;
-      }
-      const response: Response = { description: responseDetails.description || `${statusCode} response` };
-      if (responseDetails.bodyType) {
-        response.schema = { $ref: `#/definitions/${responseDetails.bodyType}` };
-      }
-
-      formatted[statusCode] = response;
-    });
-
-    return formatted;
-  };
-
-  // httpEventToSecurity = (http: EitherHttpEvent) => {
-  //   // TODO - add security sections
-  //   return undefined
-  // }
-
-  pathToParam = (pathParam: string, paramInfoOrRequired?: PathParameterPath[string]): Parameter => {
-    const isObj = typeof paramInfoOrRequired === 'object';
-    const required = (isObj ? paramInfoOrRequired.required : paramInfoOrRequired) ?? true;
-
-    return {
-      name: pathParam,
-      in: 'path',
-      required,
-      description: isObj ? paramInfoOrRequired.description : undefined,
-      type: 'string',
-    };
-  };
-
-  // The arg is actually type `HttpEvent | HttpApiEvent`, but we only use it if it has httpEvent props (or shared props),
-  //  so we can lie to the compiler to make typing simpler
-  httpEventToParameters = (httpEvent: CustomHttpEvent): Parameter[] => {
-    const parameters: Parameter[] = [];
-
-    if (httpEvent.bodyType) {
-      parameters.push({
-        in: 'body',
-        name: 'body',
-        description: 'Body required in the request',
-        required: true,
-        schema: { $ref: `#/definitions/${httpEvent.bodyType}` },
-      });
-    }
-
-    const rawPathParams: PathParameters['path'] = httpEvent.request?.parameters?.paths;
-    const match = httpEvent.path.match(/[^{}]+(?=})/g);
-    let pathParameters: string[] = match ?? [];
-
-    if (rawPathParams) {
-      Object.entries(rawPathParams ?? {}).forEach(([param, paramInfo]) => {
-        parameters.push(this.pathToParam(param, paramInfo));
-        pathParameters = removeStringFromArray(pathParameters, param);
-      });
-    }
-
-    // If no match, will just be [] anyway
-    pathParameters.forEach((param: string) => parameters.push(this.pathToParam(param)));
-
-    if (httpEvent.headerParameters || httpEvent.request?.parameters?.headers) {
-      // If no headerParameters are provided, try to use the builtin headers
-      const rawHeaderParams: HeaderParameters =
-        httpEvent.headerParameters ??
-        Object.entries(httpEvent.request!.parameters!.headers!).reduce(
-          (acc, [name, required]) => ({ ...acc, [name]: { required, type: 'string' } }),
-          {}
-        );
-
-      Object.entries(rawHeaderParams).forEach(([param, data]) => {
-        parameters.push({
-          in: 'header',
-          name: param,
-          required: data.required ?? false,
-          type: data.type ?? 'string',
-          description: data.description,
-        });
-      });
-    }
-
-    if (httpEvent.queryStringParameters || httpEvent.request?.parameters?.querystrings) {
-      // If no queryStringParameters are provided, try to use the builtin query strings
-      const rawQueryParams: QueryStringParameters =
-        httpEvent.queryStringParameters ??
-        Object.entries(httpEvent.request!.parameters!.querystrings!).reduce(
-          (acc, [name, required]) => ({ ...acc, [name]: { required, type: 'string' } }),
-          {}
-        );
-
-      Object.entries(rawQueryParams).forEach(([param, data]) => {
-        parameters.push({
-          in: 'query',
-          name: param,
-          type: data.type ?? 'string',
-          description: data.description,
-          required: data.required ?? false,
-          ...(data.type === 'array'
-            ? {
-                items: { type: data.arrayItemsType },
-                collectionFormat: 'multi',
-              }
-            : {}),
-        });
-      });
-    }
-
-    return parameters;
   };
 }
